Background Tasks Complete V1
The process of tying pages, components, providers and the database together has been a success.

There is now a test page, test component, and test provider that all interact with each other successfully.

These each have a ton of comments in each respective section to illustrate what is happening so that the logic can be replicated in real portions of the program.

All variable names, methods, etc. have been cleaned up and clutter has  been brought to a minimum to close out version one of this branch.