Calendar Modifications
Added a button to the calendar page to navigate back and forth without having to restart, really helps in the emulator so you dont have to close and reopen the app each time.

Also I wanted to see if the script I made to automatically build the IOS and Android projects worked and reflected in Github, which they do apparently.

As of this commit everything functions on both emulators for IOS and Android, as well as browser.