New Calendar
I made the realization that the @native-ionic calendar is just an import of the methods which allow you to interact with the native device.

THe ionic-3 calendar is the one which as all the <ion> html decorators that are used to display the calendar.

So I seperated the portions of code and put a basic calendar on screen to begin working on it.