Fixed Event Add
In our efforts to plow through getting the observable to work we encountered a lot of small errors that seemed inconsequential in respect to what we were doing.

However, as I progressed it became more and more arduous as all the small errors were stacking up. It came to a point where simple functionalities were not working.

So I went backwards to start fixing the small things, in an effort to alleviate the bigger problems.

<> Finished Tasks:
1. Fixed date time pickers in the html for adding events.

2. Added the startTime, and endTime in the "add event" page's typescript

2.1. Seperated the date and the time, and stored them into seperate variables. So that now the service is pushing a "startDate, endDate, startTime, endTime" to the database as seperate fields instead of concatenated together

3. Modified the on day select to add a leading 0 to any single digit days or months coming from the $event variable. This uses an "if" structure to dynamically modify the value of the "selectedDay" accordingly.

-- These changes should yield much better results moving forward on bigger tasks as the data in the database is much more consistant and reliable.