Progress Update
My freaking Lord!

A semicolon!!!!
I had to implement an entire keydown search functionality to see what was different between the two methods and providers that was causing all this fuss!

A mother freaking semicolon!

It was so much that I need to commit this version so that I do not have to backtrack this again