Zion
We have reached the gates.

We now have:
1. Two database filters:
- one is static to hold a filter which is default of nothing, allows all events to be pulled.
- one is dynamic, allows changing of the filter to decided which events are pulled.

2. Two observables. "events" is mapped to the dynamic "databaseFilter"
- "allEvents" is mapped to the static one, so it is permanently set to a list of all events in the database.

Quality of life:
I two way data bound all of the afformentioned properties to the html page so you can see them as you type. no debugger or console.log required... they are asynchronous so they change as the values change automatically.

Now we just need to break the static "allEvents[]" down to pass it into the calendar typsescript.