Need some more eyes
I'm not exactly sure why what I have is different than what is right below it... we are scratching the surface.

But 2:46 A.M. before I have to work a double starting at 11 A.M. is just too much my friend.

Feel free to check it out... its so close. I just don't understand what the fuck is wrong anymore.