Refactored School Districts
The portion which was meant to toggle the map on page entry was not working.

The arrow was being changed to "display: none" on inspection of the element after execution, but the DOM was stilly showing it.

I said fuck it and changed the entire operation to build the html elements off *ng-if directives respective to whichever district was selected at the time... as "ng-show" I believe is now depriciated. Works like a charm